SmartReporter/ai-service/requirements.txt
Flask==2.3.0
torch==2.0.0
torchvision==0.15.0
yolov5==7.0.0
opencv-python==4.7.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==9.5.0
